Tried to use modules properly, gave up and back to generating a single file.
