+++ tools/Unicode/CaseInsensitive/.gitignore
@@ -1,4 +1,5 @@
 node_modules/
+typings/
 *.js
 *.map
 *.txt
