+++ tools/Unicode/CaseInsensitive/tsconfig.json
@@ -1,8 +1,8 @@
 {
     "compilerOptions": {
         "target": "es6",
-        "module": "amd",
         "sourceMap": true,
-        "outFile": "mappings.js"
+        "outFile": "mappings.js",
+        "module": "amd"
     }
 }
